🔩: unify interfaces
